Added namespaced target aliases
When installing this library, all targets are exported with the prefix
    "Microsoft::"
However, when using this project via `add_subdirectory` all targets lack the
prefix.

Due to this, users were required to change the name by which they reference
targets exported by this project depending on the strategy they use to include
this project.

This problem is being solved by adding an alias i.e.:
    add_library(Microsoft::DirectX-Headers ALIAS DirectX-Headers)